+++ R/forest.netsplit.R
@@ -612,11 +612,13 @@ forest.netsplit <- function(x,
         suppressWarnings(metagen(dat$TE, dat$seTE,
                                  studlab = dat$evidence, data = dat,
                                  sm = x$sm,
+                                 method.tau = "DL",
                                  byvar = dat$comps, print.byvar = FALSE))
     else
       m <-
         suppressWarnings(metagen(dat$TE, dat$seTE,
-                                 studlab = dat$comps, data = dat, sm = x$sm))
+                                 studlab = dat$comps, data = dat, sm = x$sm,
+                                 method.tau = "DL"))
     ##
     if (overall) {
       m$w.fixed[m$studlab == text.overall] <- max(m$w.fixed, na.rm = TRUE)
@@ -667,11 +669,13 @@ forest.netsplit <- function(x,
         suppressWarnings(metagen(dat$TE, dat$seTE,
                                  studlab = dat$comps, data = dat,
                                  sm = x$sm,
+                                 method.tau = "DL",
                                  byvar = dat$evidence, print.byvar = FALSE))
     else
       m <-
         suppressWarnings(metagen(dat$TE, dat$seTE,
-                                 studlab = dat$comps, data = dat, sm = x$sm))
+                                 studlab = dat$comps, data = dat, sm = x$sm,
+                                 method.tau = "DL"))
     ##
     if (overall) {
       m$w.fixed[m$byvar == text.overall] <- max(m$w.fixed, na.rm = TRUE)