+++ R/funnel.netmeta.R
@@ -331,7 +331,8 @@ funnel.netmeta <- function(x,
   ##
   m.adj <-
     suppressWarnings(metagen(res$TE.adj, res$seTE,
-                             studlab = res$studlab, sm = x$sm))
+                             studlab = res$studlab, sm = x$sm,
+                             method.tau = "DL"))
   ##
   n.comps <- length(unique(res$comparison))
   ##