+++ R/print.netcomb.R
@@ -192,9 +192,9 @@ print.netcomb <- function(x,
   ##  
   if (fixed | random) {
     cat(paste("Number of studies: k = ", x$k, "\n", sep = ""))
+    cat(paste("Number of pairwise comparisons: m = ", x$m, "\n", sep = ""))
     cat(paste("Number of treatments: n = ", x$n, "\n", sep = ""))
     cat(paste("Number of active components: c = ", x$c, "\n", sep = ""))
-    cat(paste("Number of pairwise comparisons: m = ", x$m, "\n", sep = ""))
     if (!is.null(x$d))
       cat(paste("Number of designs: d = ", x$d, "\n", sep = ""))
     if (inherits(x, "discomb"))